pyproject.toml
[project]
name = "todolist"
version = "0.1.0"
description = ""
authors = [
    {name = "Sadra",email = "Sadrafadaei@gmail.com"}
]
readme = "README.md"
packages = [{include = "todolist", from = "src"}]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
